@@ -14,53 +14,59 @@
   },
   "homepage": "https://www.balistos.com",
   "dependencies": {
-    "axios": "0.18.0",
+    "@firebase/auth-types": "^0.10.0",
+    "axios": "0.19.2",
     "create-react-class": "15.6.3",
+    "firebase": "^7.13.1",
     "grid-styled": "5.0.2",
-    "react": "16.8.4",
-    "react-dom": "16.8.4",
-    "react-redux": "6.0.1",
-    "react-router-dom": "4.4.0",
-    "react-slider": "0.11.2",
+    "prettier": "^2.0.2",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-redux": "7.2.0",
+    "react-redux-firebase": "^3.3.0",
+    "react-router-dom": "5.1.2",
+    "react-slider": "1.0.3",
     "react-youtube": "7.9.0",
-    "redux": "4.0.1",
+    "redux": "4.0.5",
     "redux-devtools-extension": "^2.13.8",
-    "redux-saga": "1.0.2",
-    "socket.io-client": "2.2.0",
-    "styled-components": "4.1.3",
-    "typescript": "3.3.3333",
+    "redux-firestore": "^0.13.0",
+    "redux-saga": "1.1.3",
+    "socket.io-client": "2.3.0",
+    "styled-components": "5.0.1",
+    "typescript": "3.8.3",
     "video-time": "1.1.0"
   },
   "devDependencies": {
-    "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
     "@types/grid-styled": "^4.2.1",
-    "@types/jest": "^24.0.11",
-    "@types/node": "^11.11.0",
-    "@types/react": "^16.8.7",
-    "@types/react-dom": "^16.8.2",
-    "@types/react-redux": "^7.0.3",
-    "@types/react-router-dom": "^4.3.1",
-    "@types/react-slider": "^0.8.1",
-    "@types/react-test-renderer": "^16.8.1",
-    "@types/react-youtube": "^7.6.0",
+    "@types/jest": "25.1.4",
+    "@types/node": "^13.9.8",
+    "@types/react": "^16.9.31",
+    "@types/react-dom": "^16.9.6",
+    "@types/react-redux": "^7.1.7",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/react-slider": "^1.0.0",
+    "@types/react-test-renderer": "^16.9.2",
+    "@types/react-youtube": "^7.6.2",
     "@types/socket.io-client": "^1.4.32",
-    "@types/styled-components": "^4.1.8",
-    "@types/webpack-env": "^1.13.7",
-    "cross-env": "5.2.0",
-    "cypress": "3.1.5",
-    "enzyme": "3.8.0",
-    "enzyme-adapter-react-16": "1.9.1",
-    "enzyme-to-json": "3.3.5",
-    "eslint-config-airbnb": "17.1.0",
-    "eslint-find-rules": "3.3.1",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-jsx-a11y": "6.2.1",
-    "eslint-plugin-react": "7.12.4",
-    "jest-styled-components": "6.3.1",
-    "react-scripts": "2.1.8",
-    "react-test-renderer": "16.8.4",
-    "tslint": "^5.13.1",
-    "tslint-react": "^3.6.0"
+    "@types/styled-components": "^5.0.1",
+    "@types/webpack-env": "^1.15.1",
+    "cross-env": "7.0.2",
+    "cypress": "4.3.0",
+    "enzyme": "3.11.0",
+    "enzyme-adapter-react-16": "1.15.2",
+    "enzyme-to-json": "3.4.4",
+    "eslint-config-airbnb": "18.1.0",
+    "eslint-find-rules": "3.4.0",
+    "eslint-plugin-import": "2.20.0",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.19.0",
+    "jest-styled-components": "7.0.0",
+    "prettier-tslint": "^0.4.2",
+    "react-scripts": "3.4.1",
+    "react-test-renderer": "16.13.1",
+    "tslint": "6.1.0",
+    "tslint-react": "4.2.0"
   },
   "scripts": {
     "start": "cross-env REACT_APP_API_ENDPOINT=production react-scripts start",
