@@ -1,95 +1,81 @@
 // @flow
 
-import React, { FormEvent } from "react";
-import styled from "styled-components";
+import React, { FormEvent } from 'react';
+import styled from 'styled-components';
 
-import Button from "../common/Button";
-import Input from "../common/Input";
-import Dropdown from "./Dropdown";
+import Button from '../common/Button';
+import Input from '../common/Input';
+import Dropdown from './Dropdown';
+import { createPlaylist } from '../../services/firestore.service';
 
 const LabelTitle = styled.div`
-  font-weight: 700;
-  font-size: 13px;
-  color: #3e414c;
-  float: left;
-  line-height: 24px;
+    font-weight: 700;
+    font-size: 13px;
+    color: #3e414c;
+    float: left;
+    line-height: 24px;
 `;
 
 const Label = styled.label`
-  margin-bottom: 5px;
-  margin-right: 20px;
-  width: 100%;
-  float: left;
-  padding-bottom: 5px;
+    margin-bottom: 5px;
+    margin-right: 20px;
+    width: 100%;
+    float: left;
+    padding-bottom: 5px;
 `;
 
 interface Props {
-  onCreatePlaylistSubmit: (title: string, description: string) => void;
+    setOpen: (window: string) => void;
 }
 
-interface State {
-  title: string;
-  description: string;
-}
+const CreatePlaylist = (props: Props) => {
+    const [form, setForm] = React.useState({ title: '', description: '' });
 
-class CreatePlaylist extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      description: "",
-      title: "",
-    };
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  public handleChange(event: FormEvent<HTMLFormElement>) {
-    const update: any = {};
-    const target = event.target as HTMLInputElement;
-    update[target.name] = target.value;
-    this.setState(update);
-  }
-
-  public handleSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    if (!this.state.title.trim() || !this.state.description.trim()) {
-      return;
+    function handleChange(event: FormEvent<HTMLFormElement>) {
+        const target = event.target as HTMLInputElement;
+        setForm({ ...form, [target.name]: target.value });
     }
-    this.props.onCreatePlaylistSubmit(this.state.title, this.state.description);
-    this.setState({
-      title: "",
-      description: "",
-    });
-  }
 
-  public render() {
+    function handleSubmit(event: FormEvent<HTMLFormElement>) {
+        event.preventDefault();
+        if (!form.title.trim() || !form.description.trim()) {
+            return;
+        }
+        createPlaylist(form.title, form.description);
+        setForm({
+            title: '',
+            description: '',
+        });
+        props.setOpen('');
+    }
+
     return (
-      <Dropdown>
-        <form onSubmit={this.handleSubmit}>
-          <Label htmlFor="title">
-            <LabelTitle>Title</LabelTitle>
-            <Input
-              type="text"
-              name="title"
-              onChange={this.handleChange}
-              value={this.state.title}
-            />
-          </Label>
-          <Label htmlFor="description">
-            <LabelTitle>Description</LabelTitle>
-            <Input
-              type="text"
-              name="description"
-              onChange={this.handleChange}
-              value={this.state.description}
-            />
-          </Label>
-          <Button type="submit" green={true}>Create</Button>
-        </form>
-      </Dropdown>
+        <Dropdown>
+            <form onSubmit={handleSubmit}>
+                <Label htmlFor="title">
+                    <LabelTitle>Title</LabelTitle>
+                    <Input
+                        type="text"
+                        name="title"
+                        onChange={handleChange}
+                        value={form.title}
+                    />
+                </Label>
+                <Label htmlFor="description">
+                    <LabelTitle>Description</LabelTitle>
+                    <Input
+                        type="text"
+                        name="description"
+                        onChange={handleChange}
+                        value={form.description}
+                    />
+                </Label>
+                <Button type="submit" green={true}>
+                    Create
+                </Button>
+            </form>
+        </Dropdown>
     );
-  }
-}
+};
 
 export default CreatePlaylist;
