@@ -1,5 +1,5 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react'
+import styled from 'styled-components'
 
 const DropdownContainer = styled.div`
     border: 1px solid #6d6d6d;
@@ -10,33 +10,30 @@ const DropdownContainer = styled.div`
     right: 5px;
     border-top: 0px;
     padding: 10px;
-    width: ${(props: DropdownProps) => (props.small ? "210px" : "302px")};
+    width: ${(props: DropdownProps) => (props.small ? '210px' : '302px')};
 
     @media (min-width: 40em) {
         right: 15px;
     }
-`;
+`
 
 interface DropdownProps {
-  small?: boolean;
+    small?: boolean
 }
 
-const Dropdown = (props: {
-    small?: boolean,
-    children: any,
-  }) => {
-  function stopPropagation(event: any) {
-    event.stopPropagation();
-  }
-  return (
-    <DropdownContainer onClick={stopPropagation} small={props.small}>
-      {props.children}
-    </DropdownContainer>
-  );
-};
+const Dropdown = (props: { small?: boolean; children: any }) => {
+    function stopPropagation(event: any) {
+        event.stopPropagation()
+    }
+    return (
+        <DropdownContainer onClick={stopPropagation} small={props.small}>
+            {props.children}
+        </DropdownContainer>
+    )
+}
 
 Dropdown.defaultProps = {
-  small: false,
-};
+    small: false,
+}
 
-export default Dropdown;
+export default Dropdown
