@@ -1,16 +1,22 @@
 // @flow
 
-import React from "react";
+import React from 'react';
 
-import Button from "../common/Button";
-import Dropdown from "./Dropdown";
+import Button from '../common/Button';
+import Dropdown from './Dropdown';
+import { firebase } from '../../config/firebase';
 
-const LogOut = (props: {
-                  onLogoutClick: () => void,
-                }) => (
-                  <Dropdown small={true}>
-                    <Button data-cy="log-out" green={true} onClick={props.onLogoutClick}>Log Out</Button>
-                  </Dropdown>
-);
+const LogOut = () => {
+    function logout() {
+        firebase.auth().signOut();
+    }
+    return (
+        <Dropdown small={true}>
+            <Button data-cy="log-out" green={true} onClick={logout}>
+                Log Out
+            </Button>
+        </Dropdown>
+    );
+};
 
 export default LogOut;
