@@ -1,60 +1,37 @@
 // @flow
 
-import { Box, Flex } from "grid-styled";
-import React from "react";
-import { connect } from "react-redux";
+import { Box, Flex } from 'grid-styled';
+import React from 'react';
+import { firestoreConnect } from 'react-redux-firebase';
+import { compose } from 'redux';
+import { connect } from 'react-redux';
 
-import { fetchPopularPlaylists } from "../../actions";
-import PopularPlaylist from "./PopularPlaylist";
-
-import { Dispatch } from "redux";
-import { PlaylistType } from "../../types";
-
-function mapStateToProps(state: any) {
-  return {
-    playlists: state.results.popular,
-  };
-}
-
-const mapDispatchToProps = (dispatch: Dispatch) => ({
-  fetchPopularPlaylists: () => {
-    dispatch(fetchPopularPlaylists());
-  },
-});
+import PopularPlaylist from './PopularPlaylist';
+import { PlaylistType } from '../../types';
 
 interface Props {
-  fetchPopularPlaylists: () => void;
-  playlists: PlaylistType[];
+    playlists: PlaylistType[];
 }
 
-interface State {
-}
-
-class PopularPlaylistContainer extends React.Component<Props, State> {
-  public componentWillMount() {
-    this.props.fetchPopularPlaylists();
-  }
-  public render() {
+const PopularPlaylistContainer: React.FC<Props> = (props: Props) => {
     return (
-      <Flex flexWrap={"wrap"}>
-        <Box width={[1, 1, 1 / 2, 1 / 2]} px={2}>
-          {this.props.playlists
-            .filter((playlist, index) => index < this.props.playlists.length / 2)
-            .map((result, index) =>
-              <PopularPlaylist data={result} index={index} key={result.id} />)}
-        </Box>
-        <Box width={[1, 1, 1 / 2, 1 / 2]} px={2}>
-          {this.props.playlists
-            .filter((playlist, index) => index >= this.props.playlists.length / 2)
-            .map((result, index) => (<PopularPlaylist
-              data={result}
-              index={Math.ceil(this.props.playlists.length / 2) + index}
-              key={result.id}
-            />))}
-        </Box>
-      </Flex>
+        <Flex flexWrap={'wrap'}>
+            <Box width={[1, 1, 1 / 2, 1 / 2]} px={2}>
+                {props.playlists.map((result: PlaylistType, index: number) => (
+                    <PopularPlaylist
+                        {...result}
+                        index={index}
+                        key={result.id}
+                    />
+                ))}
+            </Box>
+        </Flex>
     );
-  }
-}
+};
 
-export default connect(mapStateToProps, mapDispatchToProps)(PopularPlaylistContainer);
+export default compose(
+    firestoreConnect(() => ['playlists']),
+    connect((state: any) => ({
+        playlists: state.firestore.ordered.playlists || [],
+    }))
+)(PopularPlaylistContainer) as React.ComponentType;
