@@ -1,153 +1,133 @@
-// @flow
-
-import React, { ChangeEvent, FormEvent } from "react";
-import { connect } from "react-redux";
-import styled from "styled-components";
-
-import { Dispatch } from "redux";
-import { addVideo, clearYoutubeResults, resetYoutubeSearchQuery, searchYoutube, updateSearchIndex } from "../../actions";
-import { YoutubeResultVideoType } from "../../types/index";
-import VideoResult from "./VideoResult";
-
-const mapDispatchToProps = (dispatch: Dispatch) => ({
-  searchYoutube: (e: FormEvent<HTMLFormElement>) => {
-    dispatch(searchYoutube(e.currentTarget.value));
-  },
-  addVideo: (id: string, title: string, playlistId: string) => {
-    dispatch(addVideo(id, title, playlistId, false));
-  },
-  updateSearchIndex: (value: number) => {
-    dispatch(updateSearchIndex(value));
-  },
-  clearYoutubeResults: () => {
-    dispatch(clearYoutubeResults());
-  },
-  resetYoutubeSearchQuery: () => {
-    dispatch(resetYoutubeSearchQuery());
-  },
-});
-
-const mapStateToProps = (state: any, ownProps: any) => ({
-  ...ownProps,
-  results: state.results.youtube,
-  query: state.results.query,
-  index: state.results.youtubeIndex,
-});
+import React, { FormEvent } from 'react';
+import { connect } from 'react-redux';
+import styled from 'styled-components';
+import { AuthUserType } from '../../types/index';
+import VideoResult from './VideoResult';
+import { addVideoToPlaylist } from '../../services/firestore.service';
+import { searchYoutubeVideos } from '../../services/youtube.service';
 
 const Search = styled.div`
-  float: left;
-  position: relative;
-  width: 100%;
-  padding: 5px;
+    float: left;
+    position: relative;
+    width: 100%;
+    padding: 5px;
 `;
 
 const SearchInput = styled.input`
-  width: 100%;
-  height: 40px;
-  border-radius: 0px;
-  float: left;
-  line-height: 18px;
-  font-size: 13px;
-  color: #333333;
-  background-color: white;
-  padding: 11px 10px;
-  margin: 0px;
+    width: 100%;
+    height: 40px;
+    border-radius: 0px;
+    float: left;
+    line-height: 18px;
+    font-size: 13px;
+    color: #333333;
+    background-color: white;
+    padding: 11px 10px;
+    margin: 0px;
 `;
 
 const SearchResults = styled.div`
-  position: absolute;
-  top: 43px;
-  width: calc(100% - 10px);
-  z-index: 20;
-  background: #F6F6F6;
-  border: 1px solid #CCCCCC;
-  border-radius: 3px;
-  list-style: none;
-  padding: 0px;
-  text-align: left;
+    position: absolute;
+    top: 43px;
+    width: calc(100% - 10px);
+    z-index: 20;
+    background: #f6f6f6;
+    border: 1px solid #cccccc;
+    border-radius: 3px;
+    list-style: none;
+    padding: 0px;
+    text-align: left;
 `;
 
 interface Props {
-  query: string;
-  addVideo: (id: string, title: string, playlistId: string) => void;
-  searchYoutube: (event: ChangeEvent<HTMLInputElement>) => void;
-  updateSearchIndex: (index: number) => void;
-  resetYoutubeSearchQuery: () => void;
-  clearYoutubeResults: () => void;
-  index: number;
-  id: string;
-  results: YoutubeResultVideoType[];
+    id: string;
+    user: AuthUserType;
 }
 
-interface State {}
+const SearchVideo = (props: Props) => {
+    const [results, setResults] = React.useState<any[]>([]);
+    const [query, setQuery] = React.useState('');
+    const [index, setIndex] = React.useState(0);
 
-class SearchVideo extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.handleKeyEvent = this.handleKeyEvent.bind(this);
-  }
-  public componentDidMount() {
-    document.addEventListener("keyup", this.handleKeyEvent, false);
-  }
-  public componentWillUnmount() {
-    document.removeEventListener("keyup", this.handleKeyEvent, false);
-  }
-
-  public handleKeyEvent(event: KeyboardEvent) {
-    const key = event.key;
-    let index;
-    switch (key) {
-      case "ArrowUp":
-        this.props.updateSearchIndex(-1);
-        break;
-      case "ArrowDown":
-        this.props.updateSearchIndex(1);
-        break;
-      case "Enter":
-        index = this.props.index < 0 ? ((5 + this.props.index) % 5) : this.props.index % 5;
-        if (this.props.results[index]) {
-          this.props.addVideo(
-            this.props.results[index].id.videoId,
-            this.props.results[index].snippet.title, this.props.id,
-          );
-        }
-        break;
-      case "Escape":
-        this.props.clearYoutubeResults();
-        this.props.resetYoutubeSearchQuery();
-        break;
-      default:
-        break;
-    }
-  }
-  public render() {
-    return (
-      <Search>
-        <SearchInput
-          type="text"
-          placeholder="Search for YouTube video and add to playlist"
-          onChange={this.props.searchYoutube}
-          value={this.props.query || ""}
-          autoComplete="off"
-        />
-        {this.props.results && this.props.query ?
-          <SearchResults>
-            {this.props.results.map((result, index) =>
-              (<VideoResult
-                title={result.snippet.title}
-                image={result.snippet.thumbnails.default.url}
-                onItemClick={() => this.props.addVideo(
-                  result.id.videoId,
-                  result.snippet.title, this.props.id,
-                )}
-                id={result.id.videoId}
-                key={result.id.videoId}
-                active={index === this.props.index % 5 || index === ((5 + this.props.index) % 5)}
-              />))}
-          </SearchResults> : undefined }
-      </Search>
+    const addVideo = React.useCallback(
+        (videoId: string, title: string) => {
+            addVideoToPlaylist(props.id, videoId, title);
+        },
+        [props.id]
     );
-  }
-}
 
-export default connect(mapStateToProps, mapDispatchToProps)(SearchVideo);
+    const handleKeyEvent = React.useCallback(
+        (event: KeyboardEvent) => {
+            const key = event.key;
+            switch (key) {
+                case 'ArrowUp':
+                    setIndex(index - 1);
+                    break;
+                case 'ArrowDown':
+                    setIndex(index + 1);
+                    break;
+                case 'Enter':
+                    setIndex(index < 0 ? (5 + index) % 5 : index % 5);
+                    if (results[index]) {
+                        addVideo(
+                            results[index].id.videoId,
+                            results[index].snippet.title
+                        );
+                    }
+                    break;
+                case 'Escape':
+                    setResults([]);
+                    setQuery('');
+                    break;
+                default:
+                    break;
+            }
+        },
+        [addVideo, index, results]
+    );
+
+    React.useEffect(() => {
+        document.addEventListener('keyup', handleKeyEvent, false);
+        return () =>
+            document.removeEventListener('keyup', handleKeyEvent, false);
+    }, [index, results, handleKeyEvent]);
+
+    function searchYoutube(e: FormEvent<HTMLInputElement>) {
+        const target = e.target as HTMLInputElement;
+        const value = target.value;
+        setQuery(value);
+        searchYoutubeVideos(value).then((res) => {
+            setResults(res);
+        });
+    }
+
+    return (
+        <Search>
+            <SearchInput
+                type="text"
+                placeholder="Search for YouTube video and add to playlist"
+                onChange={searchYoutube}
+                value={query}
+                autoComplete="off"
+            />
+            {results && query && (
+                <SearchResults>
+                    {results.map((result: any, i: number) => (
+                        <VideoResult
+                            title={result.snippet.title}
+                            image={result.snippet.thumbnails.default.url}
+                            id={result.id.videoId}
+                            key={result.id.videoId}
+                            active={i === index % 5 || i === (5 + index) % 5}
+                            addVideo={addVideo}
+                        />
+                    ))}
+                </SearchResults>
+            )}
+        </Search>
+    );
+};
+
+export default connect((state: any) => ({
+    user: state.firebase.auth,
+}))(SearchVideo);
