@@ -1,44 +1,52 @@
-// @flow
+import React from 'react';
+import styled from 'styled-components';
 
-import React from "react";
-import styled from "styled-components";
+import VideoListItem from './VideoListItem';
 
-import VideoListItem from "./VideoListItem";
-
-import { VideoType } from "../../types";
+import { VideoType } from '../../types';
 
 const EmptyItem = styled.div`
-  color: #3E414C;
-  padding: 10px;
+    color: #3e414c;
+    padding: 10px;
 `;
 
 const List = styled.div`
-  width: 100%;
+    width: 100%;
 `;
 
 const VideoList = (props: {
-  videos: VideoType[],
-  current: VideoType,
-
+    videos: VideoType[];
+    current: VideoType;
+    id: string;
 }) => {
-  if (!props.videos.length) {
-    return <EmptyItem>Playlist is empty. Please search and add a video.</EmptyItem>;
-  }
+    if (!props.videos.length) {
+        return (
+            <EmptyItem>
+                Playlist is empty. Please search and add a video.
+            </EmptyItem>
+        );
+    }
 
-  return (
-    <List>
-      {props.current ? <VideoListItem
-        video={props.current}
-        key={props.current.id}
-        index={0}
-      /> : undefined }
-      {props.videos.sort((a, b) => {
-        const diff = b.likes.reduce((total, like) => total + like.value, 0) -
-          a.likes.reduce((total, like) => total + like.value, 0);
-        return diff === 0 ? a.id - b.id : diff;
-      }).map((video, index) => <VideoListItem video={video} key={video.id} index={index + 1} />)}
-    </List>
-  );
+    return (
+        <List>
+            {props.current && (
+                <VideoListItem
+                    video={props.current}
+                    key={props.current.id}
+                    id={props.id}
+                    index={0}
+                />
+            )}
+            {props.videos.map((video, index) => (
+                <VideoListItem
+                    video={video}
+                    key={video.id}
+                    id={props.id}
+                    index={index + 1}
+                />
+            ))}
+        </List>
+    );
 };
 
 export default VideoList;
