@@ -1,7 +1,7 @@
 // @flow
 
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+import styled from 'styled-components';
 
 const ContainerComponent = styled.div`
     margin: 0px auto;
@@ -10,17 +10,14 @@ const ContainerComponent = styled.div`
     max-width: 1250px;
     padding: 0px 5px;
     position: relative;
+    justify-content: space-between;
     @media (min-width: 40em) {
         padding: 0px 15px;
     }
 `;
 
-const Container = (props: {
-  children: any,
-}) => (
-  <ContainerComponent>
-    {props.children}
-  </ContainerComponent>
+const Container = (props: { children: any }) => (
+    <ContainerComponent>{props.children}</ContainerComponent>
 );
 
 export default Container;
